requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
mysql-connector-python
python-dotenv
openpyxl